Objeto

+++ src/main/java/com/jrancsar/genesis/repositories/asociacion/CargoAdministrativoRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.jrancsar.genesis.entities.asociacion.CargoAdministrativo;
 
-public interface CargoAdministrativoRepository extends JpaRepository<CargoAdministrativo, Long>{	
+public interface CargoAdministrativoRepository extends JpaRepository<CargoAdministrativo, Integer>{	
 }
