+++ src/main/java/com/jrancsar/genesis/repositories/seguridad/PerfilRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.jrancsar.genesis.entities.seguridad.Perfil;
 
-public interface PerfilRepository extends JpaRepository<Perfil, Long>{	
+public interface PerfilRepository extends JpaRepository<Perfil, Integer>{	
 }
